Moved around a lot of files and definitions, hopefully all the colors and remove_chartjunk stuff works
